Time: 0 ms (100%), Space: 8.5 MB (53.71%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -6,10 +6,11 @@ public:
         long long digit = x % 10;
         reversed = reversed * 10 + digit;
         x/= 10;
-    }  if (reversed < INT32_MIN || reversed > INT32_MAX)
+    }  
+    if (reversed < INT32_MIN || reversed > INT32_MAX)
         return 0;
 
-    return static_cast<int>(reversed);
+    return reversed;
    
     }
 };
